move the handlePayment
