+++ client/src/components/Confirmation.jsx
@@ -19,7 +19,7 @@ function Confirmation() {
             <div className="bg-white p-8 rounded-lg shadow-md text-center">
                 <h1 className="text-2xl mb-4">Tack för ditt köp</h1>
                 <button onClick={toFirstPage} className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">
-                    Bak till hemsidan!
+                    Startsida!
                 </button>
             </div>
         </div>
